Use autoboxing/autounboxing and for each loops where possible
Now that we're using Java 5, we can make the code look a little bit nicer.
Thanks to IntelliJ for inspecting the code for me.

+++ src/tiled/mapeditor/dialogs/ResizeDialog.java
@@ -191,10 +191,10 @@ public class ResizeDialog extends JDialog implements ActionListener,
         double zoom = orient.getZoom();
 
         if (evt.getPropertyName().equalsIgnoreCase("offsetX")) {
-            int val = ((Integer)evt.getNewValue()).intValue();
+            int val = (Integer) evt.getNewValue();
             offsetX.setValue((int)(val / (currentMap.getTileWidth() * zoom)));
         } else if(evt.getPropertyName().equalsIgnoreCase("offsetY")) {
-            int val = ((Integer)evt.getNewValue()).intValue();
+            int val = (Integer) evt.getNewValue();
             offsetY.setValue((int)(val / (currentMap.getTileHeight() * zoom)));
         }
     }
